requirements.txt
PyPDF2
sandyie_read
scikit-learn
nltk
sumy
transformers
torch
bert-extractive-summarizer